+++ src/controller/consumer/get_contracts_controller.js
@@ -1,12 +1,20 @@
+'use strict';
 
-const { getContracts } = require('../service');
+const { getContracts } = require('../../service');
+
 module.exports = async (req, res, next) => {
   try {
     const { id: profileId, type: profileType } = req.profile;
     const contracts = await getContracts({ profileId, profileType });
+    // Check for empty response
     if (!contracts || contracts.length == 0) {
-      return res.status(404).end();
+      console.log(
+          'GET Contracts, completed with status code ', 404);
+      return res.status(404).send({
+        message: 'No contract found' });
     }
+    console.log(
+        'GET Contracts, completed with status code ', 200);
     res.status(200).send({
       status: 'ok',
       result: { contracts }
