+++ src/controller/consumer/get_unpdaid_jobs_controller.js
@@ -1,12 +1,20 @@
+'use strict';
 
-const { getUnpaidJobs } = require('../service');
+const { getUnpaidJobs } = require('../../service');
+
 module.exports = async (req, res, next) => {
   try {
     const { id: profileId, type: profileType } = req.profile;
     const jobs = await getUnpaidJobs({ profileId, profileType });
+    // Check for empty response
     if (!jobs || jobs.length == 0) {
-      return res.status(404).end();
+      console.log(
+          'GET Unpaid Jobs, completed with status code ', 404);
+      return res.status(404).send({
+        message: 'Job not found' });
     }
+    console.log(
+        'GET Unpaid Jobs, completed with status code ', 200);
     res.status(200).send({
       status: 'ok',
       result: { jobs }
