+++ src/database/index.js
@@ -1,3 +1,4 @@
+'use strict';
 
 const { Sequelize } = require('sequelize');
