+++ src/middleware/index.js
@@ -1,4 +1,6 @@
 
+'use strict';
+
 module.exports = {
   getProfile: require('./get_profile_middleware')
 };
