+++ tests/e2e/admin/get_admin_best_profession.test.js
@@ -1,5 +1,7 @@
+'use strict';
+
 const request = require('supertest');
-const app = require('../../src/app');
+const app = require('../../../app');
 
 describe('GET /admin/best-profession', () => {
   it('returns 404 on no Job data found', async () => {
