+++ tests/e2e/consumer/get_unpaid_jobs.test.js
@@ -1,5 +1,7 @@
+'use strict';
+
 const request = require('supertest');
-const app = require('../../src/app');
+const app = require('../../../app');
 
 describe('GET /jobs/unpaid', () => {
   it('returns 404 on no unpaid jobs', async () => {
