+++ tests/e2e/consumer/post_job_payment.test.js
@@ -1,5 +1,7 @@
+'use strict';
+
 const request = require('supertest');
-const app = require('../../src/app');
+const app = require('../../../app');
 
 describe('POST /jobs/:job_id/pay', () => {
   it('returns 401 on invalid user', async () => {
